+++ Magic/CHANGELOG.md
@@ -50,6 +50,7 @@
  - Fix warps getting corrupted if the target world is not loaded at server start
  - Fix secure/lockpick spells when non-opped
  - Fix having multiple crafting recipes with similar ingredients
+ - Fix using `thickness` parameter with Volume action
  - The `/mwarp go` and `/send` commands will try to load the target world if it is not loaded
  - Add a third alt-cast spell to wands, for a total of 4 possible fixed spells
  - Add SuperWalls spells to default configs